Add support for running external binaries
When the app is started, it will copy the `sample_arm_bin` in the
`assets` directory to the data directory (since Android mounts app data
as noexec or something) and execute it, displaying the output.
Currently, `sample_arm_bin` is an ARM ELF executable file that prints
'Hello world!' and exits.